camera/all_in_one/requirements.txt
Flask
RPi.Gpio
zeroconf
